+++ ai-services/assets/catalog/podman/values.yaml
@@ -4,7 +4,7 @@ ui:
 
 backend:
   port: ""
-  image: icr.io/ai-services-cicd/ai-services:v0.0.124
+  image: icr.io/ai-services-cicd/ai-services:v0.0.125
   runtime: ""
   adminPasswordHash: ""
   podman:
